pyproject.toml
[tool.poetry]
name = "AG1_AEtherBus"
version = "0.1.1"
description = "Redis-based async message bus for agent communication (envelope-based, reusable, production-ready)"
authors = ["Sean C. <sean@propensity.co.za>"]
license = "MIT"
keywords = ["redis", "asyncio", "message bus", "envelope", "agent", "python"]
homepage = "https://github.com/agentic1/AG1_AEtherBus"
repository = "https://github.com/agentic1/AG1_AEtherBus"
documentation = "https://github.com/agentic1/AG1_AEtherBus#readme"
readme = "README.md"
packages = [
    { include = "AG1_AetherBus" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
redis = "5.2.1"
aiohttp = "*"
python-dotenv = "*"
aiogram = "^3.20.0.post0"
aiohttp-cors = "^0.8.1"
websockets = "^15.0.1"
mcp = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
